restore all tests (still needs to add new tests)
