+++ test/entity-manager/on-logic.test.js
@@ -18,7 +18,7 @@ describe('EntityManager', function() {
             this.entities.push({ id: 1, entity: { components: this.position | this.velocity | this.clickable } })
             this.entities.push({ id: 2, entity: { components: this.position | this.velocity | this.spawnable } })
             
-            this.entityManager.getEntities = sinon.stub().returns(this.entities)
+            this.entityManager.getEntitiesByIds = sinon.stub().returns(this.entities)
             
             this.logicSpy1 = sinon.spy()
             this.entityManager.systemManager.logicSystems.set(1, {
@@ -74,16 +74,16 @@ describe('EntityManager', function() {
             expect(this.renderSpy2.calledOnce).to.be.false
         })
         
-        test('invokes [entityManager.getEntities()] with the correct paramters', () => {
+        test('invokes [entityManager.getEntitiesByIds()] with the correct paramters', () => {
             this.entityManager.onLogic(this.opts)
             
-            expect(this.entityManager.getEntities.calledTwice).to.be.true
+            expect(this.entityManager.getEntitiesByIds.calledTwice).to.be.true
             
             let logicSystem1 = this.entityManager.systemManager.logicSystems.get(1)
             let logicSystem2 = this.entityManager.systemManager.logicSystems.get(2)
             
-            expect(this.entityManager.getEntities.calledWith(logicSystem1.components)).to.be.true
-            expect(this.entityManager.getEntities.calledWith(logicSystem2.components)).to.be.true
+            expect(this.entityManager.getEntitiesByIds.calledWith(logicSystem1.entities)).to.be.true
+            expect(this.entityManager.getEntitiesByIds.calledWith(logicSystem2.entities)).to.be.true
         })
     })
 })
