+++ test/entity-manager/proxy-methods.test.js
@@ -175,91 +175,6 @@ describe('EntityManager', function() {
                 expect(() => this.entityManager.create(1)).to.not.throw(Error, msg)
             })
         })
-         
-        describe('registerSystem(callback, components, type = SystemType.Logic)', () => {
-            beforeEach(() => {
-                this.entityId1 = 0
-                this.entity1 = {
-                    components: this.position | this.velocity | this.stats,
-                }
-
-                this.entityId2 = 1
-                this.entity2 = {
-                    components: this.position | this.stats,
-                }
-
-                this.entityId3 = 2
-                this.entity3 = {
-                    components: this.position | this.velocity | this.stats,
-                }
-
-                this.entityId4 = 3
-                this.entity4 = {
-                    components: this.position,
-                }
-
-                this.entityId5 = 4
-                this.entity5 = {
-                    components: 0,
-                }
-
-
-                this.entityManager.entities[this.entityId1] = this.entity1
-                this.entityManager.entities[this.entityId2] = this.entity2
-                this.entityManager.entities[this.entityId3] = this.entity3
-                this.entityManager.entities[this.entityId4] = this.entity4
-                this.entityManager.entities[this.entityId5] = this.entity5
-
-                this.entityManager.currentMaxEntity = 5
-            })
-
-            test('is a function', () => {
-                expect(this.entityManager.registerSystem).to.be.a('function')
-            })
-            
-            test('invokes [systemManager.registerSystem] with the correct parameters', () => {
-                const type       = SystemType.Render
-                const components = this.position | this.velocity | this.stats
-                
-                const callback = function(entities, { delta }) { 
-                    for (const { entity } of entities) {
-                        entity.position += entity.velocity * delta
-                    }
-                }
-                
-                const spy = sinon.spy(this.entityManager.systemManager, 'registerSystem')
-                
-                const systemId = this.entityManager.registerSystem(type, components, callback)
-                
-                const entityIds = [this.entityId1, this.entityId3]
-
-                expect(spy.calledOnce).to.be.true
-                expect(spy.calledWith(type, components, entityIds, callback)).to.be.true
-                expect(systemId).to.equal(1)
-            })
-            
-            test('invokes [systemManager.registerSystem] with the correct component ids, given component names', () => {
-                const type           = SystemType.Render
-                const components     = this.position | this.velocity | this.stats
-                const componentNames = [ this.positionName, this.velocityName, this.statsName ]
-                
-                const callback = function(entities, { delta }) { 
-                    for (const { entity } of entities) {
-                        entity.position += entity.velocity * delta
-                    }
-                }
-                
-                const spy = sinon.spy(this.entityManager.systemManager, 'registerSystem')
-                
-                const systemId = this.entityManager.registerSystem(type, componentNames, callback)
-
-                const entityIds = [this.entityId1, this.entityId3]
-
-                expect(spy.calledOnce).to.be.true
-                expect(spy.calledWith(type, components, entityIds, callback)).to.be.true
-                expect(systemId).to.equal(1)
-            })
-        })
         
         describe('removeSystem(systemId)', () => {
             test('is a function', () => {
